refactor: Problem 객체 파일 분리
