+++ lib/create/data/mapper/problem_mapper.dart
@@ -1,5 +1,20 @@
 import 'package:mongo_ai/create/data/dto/problem_dto.dart';
-import 'package:mongo_ai/create/presentation/create_template/controller/create_template_state.dart';
+import 'package:mongo_ai/create/domain/model/problem.dart';
+
+extension ProblemMapper on ProblemDto {
+  Problem toProblem() {
+    return Problem(
+      problemType: problemLabel,
+      question: question,
+      passage: passage,
+      options: options,
+      number: 0,
+      promptDetail: '',
+      requestContent: '',
+      cleanText: '',
+    );
+  }
+}
 
 extension ProblemDtoMapper on Problem {
   ProblemDto toProblemDto(String workbookId) {
