+++ lib/create/presentation/create_complete/controller/create_complete_state.dart
@@ -1,6 +1,6 @@
 import 'dart:typed_data';
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:mongo_ai/create/presentation/create_template/controller/create_template_state.dart';
+import 'package:mongo_ai/create/domain/model/problem.dart';
 
 part 'create_complete_state.freezed.dart';
