+++ lib/create/presentation/create_complete/controller/create_complete_view_model.dart
@@ -10,8 +10,8 @@ import 'package:mongo_ai/core/state/current_folder_id_state.dart';
 import 'package:mongo_ai/core/state/current_team_id_state.dart';
 import 'package:mongo_ai/core/utils/app_dialog.dart';
 import 'package:mongo_ai/create/domain/model/create_complete_params.dart';
+import 'package:mongo_ai/create/domain/model/problem.dart';
 import 'package:mongo_ai/create/presentation/create_complete/controller/create_complete_state.dart';
-import 'package:mongo_ai/create/presentation/create_template/controller/create_template_state.dart';
 import 'package:mongo_ai/dashboard/data/dto/workbook_table_dto.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
