+++ lib/create/presentation/create_template/widget/create_problem_list_widget.dart
@@ -3,7 +3,7 @@ import 'package:gap/gap.dart';
 
 import 'package:mongo_ai/core/style/app_color.dart';
 import 'package:mongo_ai/core/style/app_text_style.dart';
-import 'package:mongo_ai/create/presentation/create_template/controller/create_template_state.dart';
+import 'package:mongo_ai/create/domain/model/problem.dart';
 import 'package:mongo_ai/create/presentation/create_template/widget/problem_card_widget.dart';
 
 class CreateProblemListWidget extends StatefulWidget {
